+++ src/Graphics/Luminance/Core/Shader/Stage.hs
@@ -14,10 +14,12 @@
 module Graphics.Luminance.Core.Shader.Stage where
 
 import Control.Applicative ( liftA2 )
+import Control.Monad ( unless )
 import Control.Monad.Except ( MonadError(throwError) )
 import Control.Monad.IO.Class ( MonadIO(..) )
 import Control.Monad.Trans.Resource ( MonadResource, register )
 import Graphics.GL
+import Graphics.Luminance.Core.Query ( getGLExtensions )
 import Foreign.C.String ( peekCString, withCString )
 import Foreign.Marshal.Alloc ( alloca )
 import Foreign.Marshal.Array ( allocaArray )
@@ -31,29 +33,38 @@ import Foreign.Storable ( peek )
 -- |A shader 'Stage'.
 newtype Stage = Stage { stageID :: GLuint }
 
+-- |A shader 'Stage' type.
+data StageType
+  = TessControlShader
+  | TessEvaluationShader
+  | VertexShader
+  | GeometryShader
+  | FragmentShader
+    deriving (Eq,Show)
+
 -- |Create a new tessellation control shader from a 'String' representation of its source code.
 createTessCtrlShader :: (HasStageError e,MonadError e m,MonadIO m,MonadResource m) => String -> m Stage
 createTessCtrlShader = mkShader GL_TESS_CONTROL_SHADER
 
--- |Create a new tessellation evaluation shader from a 'String' representation of its source code.
-createTessEvalShader :: (HasStageError e,MonadError e m,MonadIO m,MonadResource m) => String -> m Stage
-createTessEvalShader = mkShader GL_TESS_EVALUATION_SHADER
-
--- |Create a new vertex shader from a 'String' representation of its source code.
-createVertexShader :: (HasStageError e,MonadError e m,MonadIO m,MonadResource m) => String -> m Stage
-createVertexShader = mkShader GL_VERTEX_SHADER
-
--- |Create a new geometry shader from a 'String' representation of its source code.
-createGeometryShader :: (HasStageError e,MonadError e m,MonadIO m,MonadResource m) => String -> m Stage
-createGeometryShader = mkShader GL_GEOMETRY_SHADER
-
--- |Create a new fragment shader from a 'String' representation of its source code.
-createFragmentShader :: (HasStageError e,MonadError e m,MonadIO m,MonadResource m) => String -> m Stage
-createFragmentShader = mkShader GL_FRAGMENT_SHADER
-
--- |Create a new compute shader from a 'String' representation of its source code.
-createComputeShader :: (HasStageError e,MonadError e m,MonadIO m,MonadResource m) => String -> m Stage
-createComputeShader = mkShader GL_COMPUTE_SHADER
+-- |Create a shader stage from a 'String' representation of its source code and its type.
+--
+-- Note: on some hardware and backends, /tessellation shaders/ aren’t available. That function
+-- throws 'UnsupportedStage' error in such cases.
+createStage :: (HasStageError e,MonadError e m,MonadIO m,MonadResource m)
+            => StageType
+            -> String
+            -> m Stage
+createStage t src = case t of
+    TessControlShader -> checkTessSupport "tessellation control" >> mkShader GL_TESS_CONTROL_SHADER src
+    TessEvaluationShader -> checkTessSupport "tessellation evaluation" >> mkShader GL_TESS_EVALUATION_SHADER src
+    VertexShader -> mkShader GL_VERTEX_SHADER src
+    GeometryShader -> mkShader GL_GEOMETRY_SHADER src
+    FragmentShader -> mkShader GL_FRAGMENT_SHADER src
+  where
+    checkTessSupport stage = do
+      exts <- getGLExtensions
+      unless ("GL_ARB_tessellation_shader" `elem` exts) $
+        throwError $ fromStageError (UnsupportedStage stage)
 
 -- Create a shader from the kind of shader and its source code 'String' representation.
 mkShader :: (HasStageError e,MonadError e m,MonadIO m,MonadResource m)
@@ -114,7 +125,13 @@ prependGLSLPragma src =
 --
 -- 'CompilationFailed reason' occurs when a shader fails to compile, and the 'String' 'reason'
 -- contains a description of the failure.
-newtype StageError = CompilationFailed String deriving (Eq,Show)
+--
+-- 'UnsupportedStage stage' occurs when you try to create a shader which type is not supported on
+-- the current hardware.
+data StageError
+  = CompilationFailed String 
+  | UnsupportedStage String
+    deriving (Eq,Show)
 
 -- |Types that can handle 'StageError'.
 class HasStageError a where